Tested an api with pictures
Connected to an api to get dog breeds and pictures
associated with the dog breed. Made the functionality of
said api work. Still has an issue finishing the image get.
This is a fairly big update and involves a lot of code. Note
that there is still no html nor css or images from the site
itself, this is just a push with my test code to see if I
could take this project in this general direction.